+++ src/Ui5/AbstractUi5App.php
@@ -6,12 +6,13 @@ use Flat3\Lodata\Endpoint;
 use LaravelUi5\Core\Introspection\App\Ui5AppSource;
 use LaravelUi5\Core\Ui5\Contracts\Ui5AppInterface;
 use LaravelUi5\Core\Ui5\Contracts\Ui5ModuleInterface;
+use LaravelUi5\Core\Ui5CoreServiceProvider;
 
 abstract class AbstractUi5App extends Endpoint implements Ui5AppInterface
 {
     public function __construct(protected Ui5ModuleInterface $module)
     {
-        parent::__construct($module->getName());
+        parent::__construct(Ui5CoreServiceProvider::ODATA_ROUTE_PREFIX . '/' . $module->getName());
     }
 
     public function getModule(): Ui5ModuleInterface